Validate rule and group name
Note that because the communication protocol between the client and the
daemon is so simple, these values need to be restricted. i.e. They must
not contain spaces or newlines. They probably don't need to be THIS
restricted, so another option would be just to prohibit Unicode space
characters of any kind.