Fix bugs where metadata was incorrectly loaded and/or stored
When the message was successfully retrieved from cache, storeID was
blank resulting in tags being stored to SQLite3 with an empty string
for message_id, and also resulting in the tags sometimes not being
loaded.

When the message metaedata was loaded from SQLite3, if it did get the
correct storeID then it would die because of trying to store keys into
a nil map (tags).