Clean Unix domain socket on shutdown, not startup
If you delete the file on startup, it allows a subsequently started
daemon to silently override a previously started daemon, creating all
manner of confusion. It is preferable that the second instance die
with an error message if a first instance is running.

Although this means a crashed daemon will require manual clean up, in
the long term as the daemon gets more stable this will become less and
less of an issue.